refactor bundlers so that they return a bundle
A bundle represents the list of resources to be added to the target phar. This allows us to decouple the bundlers from actually knowing the target, and keeps their single responsibility to collecting the resources. Likewise, the responsibility for adding resources stays hidden in the target phar. Additionally, the PharComposer only needs to translate a package into a bundle, but doesn't need to know that actually bundlers are involved.